+++ core/base/Module.php
@@ -3,11 +3,9 @@
 namespace luya\base;
 
 use Yii;
-use yii\helpers\Inflector;
-use yii\base\InvalidParamException;
 use yii\base\InvalidConfigException;
-use luya\helpers\FileHelper;
 use luya\console\interfaces\ImportControllerInterface;
+use luya\helpers\ObjectHelper;
 
 /**
  * LUYA Module base class.
@@ -295,29 +293,7 @@ abstract class Module extends \yii\base\Module
      */
     public function getControllerFiles()
     {
-        $files = [];
-        try { // https://github.com/yiisoft/yii2/blob/master/framework/base/Module.php#L253
-            foreach (FileHelper::findFiles($this->controllerPath) as $file) {
-                $files[$this->fileToName($this->controllerPath, $file)] = $file;
-            }
-        } catch (InvalidParamException $e) {
-            try {
-                $staticPath = static::staticBasePath() . DIRECTORY_SEPARATOR . 'controllers';
-                foreach (FileHelper::findFiles($staticPath) as $file) {
-                    $files[$this->fileToName($staticPath, $file)] = $file;
-                }
-            } catch (InvalidParamException $e) {
-                // catch if folder not found.
-            }
-        };
-        
-        return $files;
-    }
-    
-    private function fileToName($prefix, $file)
-    {
-        $value = ltrim(str_replace([$prefix, 'Controller.php'], '', $file), DIRECTORY_SEPARATOR);
-        return Inflector::camel2id($value);
+        return ObjectHelper::getControllers($this);
     }
     
     /**